backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "watchfiles>=1.1.1",
]
